+++ src/Parse/ParseFile.php
@@ -164,8 +164,8 @@ class ParseFile implements Encodable
     /**
      * Internal method used when constructing a Parse File from Parse.
      *
-     * @param $name
-     * @param $url
+     * @param string $name
+     * @param string $url
      *
      * @return ParseFile
      */
@@ -229,7 +229,6 @@ class ParseFile implements Encodable
             null,
             $this->getData(),
             $useMasterKey,
-            false,
             $mimeType
         );
     }