+++ tests/Parse/ParseQueryTest.php
@@ -2602,7 +2602,6 @@ class ParseQueryTest extends TestCase
             Helper::$restKey,
             Helper::$masterKey,
             false,
-            Helper::$accountKey
         );
         ParseClient::setServerURL('http://localhost:1337', 'parse');
 
@@ -2626,8 +2625,8 @@ class ParseQueryTest extends TestCase
     {
         $query = new ParseQuery('TestObject');
         $query->_setConditions([
-            [
-                'key'  => 'value'
+            'where' => [
+                'key'   => 'value'
             ]
         ]);