+++ tests/Parse/ParseSessionFixationTest.php
@@ -18,14 +18,14 @@ class ParseSessionFixationTest extends TestCase
         ParseClient::_unsetStorage();
 
         // indicate we should not use cookies
-        ini_set("session.use_cookies", 0);
+        @ini_set("session.use_cookies", 0);
         // indicate we can use something other than cookies
-        ini_set("session.use_only_cookies", 0);
+        @ini_set("session.use_only_cookies", 0);
         // enable transparent sid support, for url based sessions
-        ini_set("session.use_trans_sid", 1);
+        @ini_set("session.use_trans_sid", 1);
         // clear cache control for session pages
-        ini_set("session.cache_limiter", "");
-        session_start();
+        @ini_set("session.cache_limiter", "");
+        @session_start();
         Helper::setUp();
     }
 
@@ -39,7 +39,7 @@ class ParseSessionFixationTest extends TestCase
 
     public static function tearDownAfterClass() : void
     {
-        session_destroy();
+        @session_destroy();
     }
 
     public function testCookieIdChangedForAnonymous()